requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
idna==3.1
multidict==5.1.0
Pillow==8.1.0
pycparser==2.20
python-dotenv==0.15.0
python-telegram-bot==13.2
pytz==2021.1
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
yarl==1.6.3
